inline add label and improve selection details
