+++ src/components/_shared/PointContextMenu/index.js
@@ -15,7 +15,6 @@ import {
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import {
   faEdit,
-  faMinusCircle,
   faCalendarAlt,
   faClock,
   faTrash,
@@ -28,9 +27,10 @@ import { useDispatch, useSelector } from 'react-redux';
 import pointsActions from 'ducks/points/actions';
 import applicationActions from 'ducks/application/actions';
 import applicationSelectors from 'ducks/application/selectors';
-import { formattedDuration } from 'components/_shared/SelectedData/SelectedDataItem/_helpers';
+
 import { useOnClickOutside } from 'hooks/useOnClickOutside';
 import LabelAs from 'components/_shared/PointContextMenu/LabelAs';
+import { formattedDuration } from 'helpers/dateTime';
 
 const PointContextMenu = ({
   pointId: id,
