--- src/components/_shared/SelectedData/SelectedDataContextMenu/SelectedDataContextMenu.module.scss
@@ -1,38 +1,0 @@
-.selectedDataContextMenu {
-  position: absolute;
-  top: 0;
-  right: 10px;
-  min-width: 240px;
-  padding: 10px;
-  background-color: #fff;
-  box-shadow: 0.5px 4px 4px rgba(0, 0, 0, 0.25);
-  border-radius: 6px;
-  z-index: 2;
-
-  li {
-    svg {
-      min-width: 30px;
-      margin-right: 5px;
-    }
-  }
-}
-
-.selectedDataContextMenuAction {
-  display: inline-flex;
-  align-items: center;
-  padding: 10px;
-  border-radius: 6px;
-  font-size: 18px;
-  width: 100%;
-  text-align: left;
-  color: #3c475b;
-
-  svg {
-    font-size: 26px;
-    margin-right: 10px;
-  }
-
-  &:hover {
-    background-color: #e5e7fa;
-  }
-}
