+++ src/components/_shared/SelectedData/_parts/SelectedDataList/SelectedDataItem/SelectedDataItem.module.scss
