+++ src/components/_shared/SelectedData/index.js
@@ -1,39 +1,20 @@
-import React, { useState } from 'react';
+import React from 'react';
 
-import SelectedDataItem from 'components/_shared/SelectedData/SelectedDataItem';
+import { selectedDataWrapper, clearFilters } from './SelectedData.module.scss';
 
-import {
-  selectedDataWrapper,
-  selectedDataHeader,
-  selectedDataHeaderInfo,
-  selectedDataList,
-  selectedDataAction,
-  selectedDataSelection,
-  clearFilters,
-} from './SelectedData.module.scss';
-
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faEllipsisV } from '@fortawesome/pro-solid-svg-icons';
-import SelectedDataContextMenu from 'components/_shared/SelectedData/SelectedDataContextMenu';
 import { useSelector, useDispatch } from 'react-redux';
-import casesSelectors from 'ducks/cases/selectors';
 import pointsSelectors from 'ducks/points/selectors';
-import applicationSelectors from 'ducks/application/selectors';
 import pointsActions from '../../../ducks/points/actions';
+import SelectedDataHeader from 'components/_shared/SelectedData/_parts/SelectedDataHeader';
+import SelectedDataList from 'components/_shared/SelectedData/_parts/SelectedDataList/SelectedDataList';
 
-const SelectedDataList = () => {
-  const [showContentMenu, setShowContentMenu] = useState(false);
-  const activeCases = useSelector(state =>
-    casesSelectors.getActiveCases(state),
-  );
+const SelectedData = () => {
   const dispatch = useDispatch();
   const points = useSelector(state => pointsSelectors.getPoints(state));
   const filteredPoints = useSelector(state =>
     pointsSelectors.getFilteredPoints(state),
   );
   const hiddenPoints = points.filter(({ hidden }) => hidden);
-  const isPublish =
-    useSelector(state => applicationSelectors.getMode(state)) === 'publish';
 
   return (
     <div className={selectedDataWrapper}>
@@ -45,42 +26,10 @@ const SelectedDataList = () => {
           Clear all filters
         </button>
       )}
-      <div className={selectedDataHeader}>
-        <h5>Selected Data</h5>
-        <div className={selectedDataHeaderInfo}>
-          {activeCases && points.length > 0 && (
-            <p className={selectedDataSelection}>
-              {filteredPoints?.length} of {points?.length}
-            </p>
-          )}
-          {!isPublish && (
-            <button
-              id="selected-data-more-menu"
-              className={selectedDataAction}
-              onClick={() => setShowContentMenu(!showContentMenu)}
-              type="button"
-            >
-              <FontAwesomeIcon icon={faEllipsisV} />
-            </button>
-          )}
-        </div>
-
-        {showContentMenu && (
-          <SelectedDataContextMenu
-            pointsLength={points.length}
-            closeAction={() => setShowContentMenu(false)}
-          />
-        )}
-      </div>
-      {filteredPoints?.length > 0 && (
-        <ul className={selectedDataList}>
-          {filteredPoints?.map(p => (
-            <SelectedDataItem key={p.pointId} {...p} />
-          ))}
-        </ul>
-      )}
+      <SelectedDataHeader />
+      <SelectedDataList />
     </div>
   );
 };
 
-export default SelectedDataList;
+export default SelectedData;
